Update code with PR suggestions
